update to latest testrunner

+++ lib/membase/helper/bucket_helper.py
@@ -139,32 +139,35 @@ class BucketOperationHelper():
                 log.error('15 seconds sleep before calling get_buckets again...')
                 time.sleep(15)
                 buckets = rest.get_buckets()
-            log.info('deleting existing buckets {0} on {1}'.format([b.name for b in buckets], serverInfo.ip))
-            for bucket in buckets:
-                log.info("remove bucket {0} ...".format(bucket.name))
-                try:
-                    status = rest.delete_bucket(bucket.name)
-                except ServerUnavailableException as e:
-                    log.error(e)
-                    log.error('5 seconds sleep before calling delete_bucket again...')
-                    time.sleep(5)
-                    status = rest.delete_bucket(bucket.name)
-                if not status:
+            if len(buckets) > 0:
+                log.info('deleting existing buckets {0} on {1}'.format([b.name for b in buckets], serverInfo.ip))
+                for bucket in buckets:
+                    log.info("remove bucket {0} ...".format(bucket.name))
                     try:
-                        BucketOperationHelper.print_dataStorage_content(servers)
-                        log.info(StatsCommon.get_stats([serverInfo], bucket.name, "timings"))
-                    except:
-                        log.error("Unable to get timings for bucket")
-                log.info('deleted bucket : {0} from {1}'.format(bucket.name, serverInfo.ip))
-                msg = 'bucket "{0}" was not deleted even after waiting for two minutes'.format(bucket.name)
-                if test_case:
-                    if not BucketOperationHelper.wait_for_bucket_deletion(bucket.name, rest, 200):
+                        status = rest.delete_bucket(bucket.name)
+                    except ServerUnavailableException as e:
+                        log.error(e)
+                        log.error('5 seconds sleep before calling delete_bucket again...')
+                        time.sleep(5)
+                        status = rest.delete_bucket(bucket.name)
+                    if not status:
                         try:
                             BucketOperationHelper.print_dataStorage_content(servers)
                             log.info(StatsCommon.get_stats([serverInfo], bucket.name, "timings"))
                         except:
                             log.error("Unable to get timings for bucket")
-                        test_case.fail(msg)
+                    log.info('deleted bucket : {0} from {1}'.format(bucket.name, serverInfo.ip))
+                    msg = 'bucket "{0}" was not deleted even after waiting for two minutes'.format(bucket.name)
+                    if test_case:
+                        if not BucketOperationHelper.wait_for_bucket_deletion(bucket.name, rest, 200):
+                            try:
+                                BucketOperationHelper.print_dataStorage_content(servers)
+                                log.info(StatsCommon.get_stats([serverInfo], bucket.name, "timings"))
+                            except:
+                                log.error("Unable to get timings for bucket")
+                            test_case.fail(msg)
+                log.info("sleep 2 seconds to make sure all buckets ({}) were deleted completely.".format([b.name for b in buckets]))
+                time.sleep(2)
 
     @staticmethod
     def delete_bucket_or_assert(serverInfo, bucket='default', test_case=None):
@@ -238,7 +241,7 @@ class BucketOperationHelper():
             if not helper.bucket_exists(bucket):
                 return True
             else:
-                time.sleep(2)
+                time.sleep(0.1)
         return False
 
     @staticmethod
@@ -253,7 +256,7 @@ class BucketOperationHelper():
             if helper.bucket_exists(bucket):
                 return True
             else:
-                time.sleep(2)
+                time.sleep(0.1)
         return False
 
     @staticmethod
@@ -278,15 +281,16 @@ class BucketOperationHelper():
         while time.time() < end_time and len(ready_vbuckets) < vbucket_count:
             for every_ip_port in server_dict:
                 #Retrieve memcached ip and port
-                ip, port = every_ip_port.split(":")
+                ip = every_ip_port.rsplit(":", 1)[0]
+                port = every_ip_port.rsplit(":", 1)[1]
                 client = MemcachedClient(ip, int(port), timeout=30)
                 client.vbucket_count = len(vbuckets)
                 bucket_info = rest.get_bucket(bucket)
-                versions = rest.get_nodes_versions(logging=False)
-                pre_spock = False
-                for version in versions:
-                    if "5" > version:
-                        pre_spock = True
+                cluster_compatibility = rest.check_cluster_compatibility("5.0")
+                if cluster_compatibility is None:
+                    pre_spock = True
+                else:
+                    pre_spock = not cluster_compatibility
                 if pre_spock:
                     log.info("Atleast 1 of the server is on pre-spock "
                              "version. Using the old ssl auth to connect to "
